negotiate as separate component
